Doubly.cpp
A doubly-linked list is a linked data structure that consists of a set of sequentially linked records called nodes. Each node contains two fields, called links, that are references to the previous and to the next node in the sequence of nodes.
This is a doubly linked list program in C++. Each Node will have a reference pointer to its next as well as previous node.

It maintains an ascending order. So new elements are inserted in the right place.

Reversing A Linked List will reverse the list as well as the property Ascending becomes descending and vice-versa. So Order of various Process :

Insertion : O(n)

Deletion : O(n)

Reversal : O(n)

Search : O(n)